Add files via upload
Simulates a noisy three neuron system.  Two neurons are coupled to each other via unidirectional excitatory synapses, the third neuron acts as an input to the first two and receives no feedback.  Each neuron is based on the Morris-Lecar model and the system is solved using the Euler-Maruyama numerical method.